move configobj.py to utils
